+++ Makefile
@@ -1,0 +1,2 @@
+init:
+	  pip install -r requirements.txt
