+++ requirements.txt
@@ -1,0 +1,4 @@
+discord
+pylyrics
+asyncio
+wikipedia
